# src/routes/application.route.js
import express from "express";
import { authUser } from "../middleware/auth.middleware.js";
import {
  handleGetAllApplications,
  handleGetApplicationById,
  handleCreateApplication,
  handleUpdateApplicationStatus,
  handleDeleteApplication,
} from "../controllers/applicatoin.controller.js";

const applicationRouter = express.Router();
applicationRouter.use(authUser);

/**
 * @swagger
 * /api/applications:
 *   get:
 *     summary: Get all applications
 *     tags: [Applications]
 *     security:
 *       - bearerAuth: []
 *     responses:
 *       200:
 *         description: List of applications
 *       401:
 *         description: Unauthorized
 */

/**
 * @swagger
 * /api/applications/{id}:
 *   get:
 *     summary: Get application by ID
 *     tags: [Applications]
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: Application details
 *       404:
 *         description: Application not found
 */

/**
 * @swagger
 * /api/applications:
 *   post:
 *     summary: Create a new application
 *     tags: [Applications]
 *     security:
 *       - bearerAuth: []
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               jobId:
 *                 type: string
 *     responses:
 *       201:
 *         description: Application created
 *       400:
 *         description: Bad request
 */

/**
 * @swagger
 * /api/applications/{id}/status:
 *   patch:
 *     summary: Update application status
 *     tags: [Applications]
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               status:
 *                 type: string
 *                 enum: [PENDING, ACCEPTED, REJECTED]
 *     responses:
 *       200:
 *         description: Application status updated
 *       404:
 *         description: Application not found
 */

/**
 * @swagger
 * /api/applications/{id}:
 *   delete:
 *     summary: Delete an application
 *     tags: [Applications]
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: Application deleted
 *       404:
 *         description: Application not found
 */

applicationRouter.get("/", handleGetAllApplications);
applicationRouter.get("/:id", handleGetApplicationById);
applicationRouter.post("/", handleCreateApplication);
applicationRouter.patch("/:id/status", handleUpdateApplicationStatus);
applicationRouter.delete("/:id", handleDeleteApplication);

export default applicationRouter;